+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/LedgerState/NewEpochState.hs
@@ -61,7 +61,7 @@ getGKeys nes = Map.keysSet $ unGenDelegs (ls ^. lsCertStateL . certDStateL . dsG
 --  contains the specified transaction outputs.
 genesisState ::
   forall era.
-  (EraGov era, EraCertState era) =>
+  (EraGov era, EraCertState era, EraStake era) =>
   Map (KeyHash 'Genesis) GenDelegPair ->
   UTxO era ->
   LedgerState era
@@ -72,7 +72,7 @@ genesisState genDelegs0 utxo0 =
         (Coin 0)
         (Coin 0)
         emptyGovState
-        (IStake mempty Map.empty)
+        mempty
         mempty
     )
     (mkCertState def def dState)