+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/State.hs
@@ -1,0 +1,8 @@
+module Cardano.Ledger.Shelley.State (
+  module Cardano.Ledger.State,
+  module Cardano.Ledger.Shelley.State.Stake,
+)
+where
+
+import Cardano.Ledger.Shelley.State.Stake
+import Cardano.Ledger.State